pyproject.toml
[project]
name = "project-manager"
version = "0.1.0"
description = "An agentic AI project manager platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.7",
    "ollama>=0.4.9",
    "python-dotenv>=1.1.0",
]
